+++ CHANGELOG.md
@@ -8,9 +8,11 @@ New:
 
 Changed:
 - Removed generated `typealias`es for package names which changed in 0.10.0.
+- In `UIViewLazyList`'s `UITableView`, adding special-case handling for programmatic scroll-to-top calls.
 
 Fixed:
-- Nothing yet!
+- Work around a problem with our memory-leak fix where our old LazyList code would crash when its placeholders were unexpectedly removed.
+- Avoid calling into the internal Zipline instance from the UI thread on startup. This would manifest as weird native crashes due to multiple threads mutating shared memory.
 
 
 ## [0.10.0] - 2024-04-05